Fix... undesirable effects on Table in slide 5
The table was weird with the border on the top on the first row, but not
one on the bottom on the last. Annoyed me as a designer. Also added
another row in HTML to act as a table head. Had to un-minify and
re-minify CSS. That's not going to be ideal for edits later.